requirements.txt
fastapi[standard]
sqlmodel
tzdata
pytest
python-jose
bcrypt
pyjwt